Refactoring: Extract InfoBox into a component
