+++ apps/judicial-system/web/src/shared-components/InfoBox/InfoBox.treat.ts
